+++ routes/admin.php
@@ -62,6 +62,10 @@ Route::middleware([Kernel::MODERATOR_ACCESS, Kernel::JSON_API])->group(function 
 
     Route::post('users/{id}/switchToAgency', [AdminController::class, 'switchUserToAgency']);
     Route::post('users/{id}/switchToRegular', [AdminController::class, 'switchUserToRegular']);
+    Route::post('users/{id}/grantAdvertising', [AdminController::class, 'grantAdvertising']);
+    Route::post('users/{id}/denyAdvertising', [AdminController::class, 'denyAdvertising']);
+    Route::post('users/{id}/grantPublishing', [AdminController::class, 'grantPublishing']);
+    Route::post('users/{id}/denyPublishing', [AdminController::class, 'denyPublishing']);
 });
 
 Route::middleware([Kernel::AGENCY_ACCESS, Kernel::JSON_API])->group(function () {